Add files via upload
This is the data analysis so far. 
For this to work, you have to put the data file that the professor sent in the same folder of this data .ipynb file. 
Similarly, pandas should be installed. I am working with anaconda so I did not have to install further libraries.